requirements.txt
# Canonical Mini Viewer pip requirements.
# Recommended install path:
#   conda env create -f env.yml
# This CUDA-capable stack can still run CPU mode with:
#   python run_viewer.py ... --device cpu --backend torch

--index-url https://download.pytorch.org/whl/cu124
--extra-index-url https://pypi.org/simple
--extra-index-url https://docs.gsplat.studio/whl/pt24cu124

# PyTorch / CUDA 12.4 runtime.
torch==2.4.1
torchvision==0.19.1

# CUDA Gaussian rasterization backend. Lazy-imported by the app.
gsplat==1.5.3+pt24cu124

# Core viewer / data / rendering stack.
numpy>=1.26,<2.3
plyfile>=1.0
pillow>=10.0
imageio[ffmpeg]>=2.34
matplotlib>=3.8
scikit-learn>=1.4
opencv-python>=4.9
viser>=1.0.15,<2
nerfview>=0.1.3
packaging>=24

# Query encoders: SigLIP2, CLIP, and DINOv2 are loaded lazily.
open-clip-torch>=2.24
transformers>=4.49,<5
sentencepiece>=0.2
protobuf>=4.25
safetensors>=0.4
huggingface-hub>=0.24
hf-xet>=1.0

# Lightweight release/dev checks.
tomli>=2.0
pyyaml>=6.0
pytest>=8
ruff>=0.6
